feat: 테스트 DisplayName 추가, 클래스 병합
